GH-41 Move method definitions to cpp file
